# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/cars/{car_id}/set_driver": {
            "post": {
                "tags": [
                    "Управление автомобилем"
                ],
                "summary": "Назначить водителя для автомобиля",
                "description": "set driver for car",
                "operationId": "setDriver",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "description": "car id",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "driver_id",
                        "in": "query",
                        "description": "driver id",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "choose driver for set",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_id",
                                    "driver_id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}